# Conflicts:
#	src/main/java/com/networkapplication/FileStorage/FileStorageManager.java
#	src/main/java/com/networkapplication/controllers/FileController.java
#	src/main/java/com/networkapplication/services/FileServiceImp.java
#	src/main/resources/application.properties

+++ src/main/java/com/networkapplication/services/FileServiceImp.java
@@ -5,6 +5,7 @@ import com.networkapplication.dtos.Request.CheckInDTO;
 import com.networkapplication.dtos.Request.FileDTORequest;
 import com.networkapplication.dtos.Response.FileDTOResponse;
 import com.networkapplication.dtos.Response.GroupFilesDTOResponse;
+import com.networkapplication.dtos.Response.ListGroupFilesDTO;
 import com.networkapplication.dtos.Response.MessageDTO;
 import com.networkapplication.exceptions.ResponseException;
 import com.networkapplication.models.File;
@@ -39,26 +40,38 @@ public class FileServiceImp implements FileService {
 
     @Override
     public MessageDTO deleteAllInGroup(Long group_id) throws ResponseException {
-
+        User user=utils.getCurrentUser();
         Group group = groupRepository.findById(group_id).orElseThrow(
                 () -> new ResponseException(404, "No Group Found")
         );
-        List<File> files = group.getFile();
-        fileRepository.deleteAll(files);
-        return MessageDTO.builder().message("Delete All Files").build();
+        if(user.getId().equals(group.getAdmin().getId()))
+        {
+            List<File> files = group.getFile();
+            fileRepository.deleteAll(files);
+            return MessageDTO.builder().message("Delete All Files").build();
+        }else {
+            throw new ResponseException(403,"unAuthorized");
+        }
     }
 
     @Override
-    public List<GroupFilesDTOResponse> loadAllGroupFiles(Long groupId) throws ResponseException {
+    public ListGroupFilesDTO loadAllGroupFiles(Long groupId) throws ResponseException {
+        User user=utils.getCurrentUser();
         Group group = groupRepository.findById(groupId).orElseThrow(
                 () -> new ResponseException(404, "Group not found"));
-        List<GroupFilesDTOResponse> filesDTOGroupResponses = new ArrayList<>();
-        for (File file :
-                group.getFile()) {
-            filesDTOGroupResponses.add(new GroupFilesDTOResponse(file));
+        if(group.getMembers().contains(user)){
+            List<GroupFilesDTOResponse> filesDTOGroupResponses = new ArrayList<>();
+            ListGroupFilesDTO listGroupFilesDTO=new ListGroupFilesDTO();
+            for (File file :
+                    group.getFile()) {
+                filesDTOGroupResponses.add(new GroupFilesDTOResponse(file));
+            }
+            listGroupFilesDTO.setGroupFilesDTOResponses(filesDTOGroupResponses);
+            return listGroupFilesDTO;
+        }else {
+            throw new ResponseException(403,"You are not in the group");
         }
 
-        return filesDTOGroupResponses;
     }
 
     @Override
@@ -67,7 +80,6 @@ public class FileServiceImp implements FileService {
         Group group = groupRepository.findById(group_id).orElseThrow(
                 () -> new ResponseException(404, "Group notFound")
         );
-
         if (group.getMembers().contains(user)) {
             for (int i = 0; i < group.getFile().size(); i++) {
                 if(Objects.equals(file.getOriginalFilename(), group.getFile().get(i).getFileName())){
@@ -79,8 +91,9 @@ public class FileServiceImp implements FileService {
             }
             File textFile = File.builder()
                     .fileName(file.getOriginalFilename())
+                    .path(uploadDirectory+"\\group"+group_id)
                     .groupFiles(group).ownerFile(user)
-                    .lastEditDate(LocalDate.of(2020, 1, 4)).build();
+                    .lastEditDate(LocalDate.now()).build();
             fileRepository.save(textFile);
             groupRepository.save(group);
             userRepository.save(user);
@@ -93,32 +106,39 @@ public class FileServiceImp implements FileService {
             bufferedOutputStream.close();
             return MessageDTO.builder().message("file upload").build();
         } else {
-            return MessageDTO.builder().message("you are not found in group").build();
+            return MessageDTO.builder().message("You are not in the group").build();
         }
     }
+
     @Override
-    public MessageDTO uploadFile(FileDTORequest fileDTORequest) throws ResponseException,IOException {
-        java.io.File folder = new java.io.File(uploadDirectory+"group "+fileDTORequest.getGroup_id());
+    public MessageDTO updateFile(MultipartFile file1,Long group_id) throws ResponseException,IOException {
+        java.io.File folder = new java.io.File(uploadDirectory+"group "+group_id);
         java.io.File[] listOfFiles = folder.listFiles();
         assert listOfFiles != null;
         User user=utils.getCurrentUser();
-        Group group = groupRepository.findById(fileDTORequest.getGroup_id()).orElseThrow(
+        Group group = groupRepository.findById(group_id).orElseThrow(
                 () -> new ResponseException(404, "Group notFound"
                 ));
         File file=null;
         if (group.getFile()!=null)
         for (int i = 0; i < group.getFile().size(); i++) {
             if (group.getFile().get(i).getCheckin().equals(user) &&
-                group.getFile().get(i).getFileName().equals(fileDTORequest.getFile().getOriginalFilename())){
+                group.getFile().get(i).getFileName().equals(file1.getOriginalFilename())){
                 file=group.getFile().get(i);
                 break;
             }
         }
         if (file==null)
             throw new ResponseException(422,"there is no such file checked in by this name");
-        for (java.io.File file1 : listOfFiles) {
-           if(fileDTORequest.getFile().getOriginalFilename().equals(file1.getName())){
-                fileStorageManager.save(fileDTORequest.getFile(),fileDTORequest.getGroup_id());
+        for (java.io.File file2 : listOfFiles) {
+           if(file1.getOriginalFilename().equals(file2.getName())){
+               String fileName = file1.getOriginalFilename();
+               boolean directory = new java.io.File(uploadDirectory+"\\group"+group_id).mkdirs();
+               java.io.File newFile = new java.io.File(uploadDirectory+"\\group"+group_id + java.io.File.separator + fileName);
+               FileOutputStream fileOutputStream = new FileOutputStream(newFile);
+               BufferedOutputStream bufferedOutputStream = new BufferedOutputStream(fileOutputStream);
+               bufferedOutputStream.write(file1.getBytes());
+               bufferedOutputStream.close();
                 return MessageDTO.builder().message("File Updated Successfully").build();
            }
         }
@@ -152,12 +172,15 @@ public class FileServiceImp implements FileService {
         return FileDTOResponse.builder().message("File Not Found In Folder").build();
     }
 
-
-
     @Override
     public synchronized MessageDTO checkIn(CheckInDTO checkIn) throws ResponseException {
         User user = utils.getCurrentUser();
+        if(checkIn.getFile_id()==null)
+        {
+            checkIn.setFile_id(List.of());
+        }
         for (int i = 0; i < checkIn.getFile_id().size(); i++) {
+
             File file = fileRepository.findById(checkIn.getFile_id().get(i)).orElseThrow(() ->
                     new ResponseException(404, "File Not Found"));
             if (file.getGroupFiles().getMembers().contains(user)) {
@@ -168,9 +191,13 @@ public class FileServiceImp implements FileService {
                 throw new ResponseException(403, "you are not found in group");
             }
         }
+        if(user.getMyFiles() == null){
+            user.setMyFiles(List.of());
+        }
         for (int j = 0; j < checkIn.getFile_id().size(); j++) {
             File file = fileRepository.findById(checkIn.getFile_id().get(j)).orElseThrow();
             file.setCheckin(user);
+            user.getMyFiles().add(file);
             fileRepository.save(file);
             userRepository.save(user);
         }
@@ -180,21 +207,28 @@ public class FileServiceImp implements FileService {
     @Override
     public MessageDTO checkOut(CheckInDTO checkOut) throws ResponseException {
         User user = utils.getCurrentUser();
-        File file = fileRepository.findById(checkOut.getFile_id().get(0)).orElseThrow(() ->
-                new ResponseException(404, "File Not Found"));
-        if (file.getGroupFiles().getMembers().contains(user)) {
-            if (file.getCheckin() != null){
-            if (file.getCheckin().getId().equals(user.getId()) ) {
-                file.setCheckin(null);
-                fileRepository.save(file);
-                userRepository.save(user);
-                return MessageDTO.builder().message("File Checked Out Successfully").build();
-            } else {
-                throw new ResponseException(401, "unAuthorized");
-            }} else {
-                throw new ResponseException(400,"File is already checked out");
-            }
-        } else
-            throw new ResponseException(403, "you are not found in group");
+        if(checkOut.getFile_id()==null)
+        {
+            checkOut.setFile_id(List.of());
+        }
+        for (int i = 0; i < checkOut.getFile_id().size(); i++) {
+            System.out.println(checkOut.getFile_id().get(i));
+            File file = fileRepository.findById(checkOut.getFile_id().get(i)).orElseThrow(() ->
+                    new ResponseException(404, "File Not Found"));
+            if (file.getGroupFiles().getMembers().contains(user)) {
+                if (file.getCheckin() != null){
+                    if (file.getCheckin().getId().equals(user.getId()) ) {
+                        file.setCheckin(null);
+                        fileRepository.save(file);
+                        userRepository.save(user);
+                    } else {
+                        throw new ResponseException(401, "unAuthorized");
+                    }} else {
+                    throw new ResponseException(400,"File is already checked out");
+                }
+            } else
+                throw new ResponseException(403, "you are not found in group");
+        }
+        return MessageDTO.builder().message("File Checked Out Successfully").build();
     }
 }
